==> notebooks/01_preprocessing.ipynb <==
# Yelp Open Dataset Preprocessing

### 1. Carga de los datasets

import json
import pandas as pd
%cd ..
from Olivia import Olivia
%cd notebooks
from utils import hypothesis_attributes

1.1. Negocios

data = []
with open("../datasets/yelp_academic_dataset_business.json", encoding = "utf-8") as data_file:
    for line in data_file:
        data.append(json.loads(line))

yelp_business = pd.DataFrame(data)
print("Total business:", yelp_business.shape[0])
yelp_business.head()

yelp_business.isna().sum()

Las horas no usarán para la demo de Oliv.ia, mientras que los atributos se asumiran con None

1.2. Reseñas

data = []
with open("../datasets/yelp_academic_dataset_review.json", encoding = "utf-8") as data_file:
    for line in data_file:
        data.append(json.loads(line))

yelp_reviews = pd.DataFrame(data)
print("Total reviews:", yelp_reviews.shape[0])
yelp_reviews.head()

yelp_reviews.isna().sum()

1.3 Fotografías

data = []
with open("../datasets/photos.json", encoding = "utf-8") as data_file:
    for line in data_file:
        data.append(json.loads(line))

yelp_photos = pd.DataFrame(data)
print("Total reviews:", yelp_photos.shape[0])
yelp_photos.head()

yelp_photos.label.unique()

### 2. Filtrado inicial de Restaurantes

import pandas as pd
from collections import Counter

yelp_business = yelp_business.dropna(subset=["categories"])
yelp_business["categories"] = yelp_business["categories"].apply(lambda x: [cat.strip() for cat in x.split(",")])
yelp_business = yelp_business[yelp_business["categories"].apply(lambda cats: "Restaurants" in cats)]
print("Total restaurants:", yelp_business.shape[0])

#### 2.1. Exploración de las categorías

categories = [category for sublist in yelp_business.categories for category in sublist]

unique_categories = set(categories)
print("Categorías únicas:", len(unique_categories))

categories = Counter(categories)
categories = categories.most_common()
filename = "categories.txt"
with open(filename, "w", encoding = "utf-8") as f:
    for category, count in categories:
        f.write(f"{category},{count}\n")
print("Categorías almacenadas en:", filename)

##### 2.1.1. Eliminación de casos inapropiados y eliminación de categorías poco informativas

## Primer filtro
# Categorías que marcan casos a excluir del dataset
categories_to_exclude = ["Flowers & Gifts", "Health & Medical", "Home Services", "Beauty & Spas", "Hair Salons", "Education", "Florists", "Tours", "Personal Chefs", "Festivals", "Specialty Schools", "Department Stores", "Amusement Parks", "Nutritionists", "Vitamins & Supplements", "Hair Salons""Religious Organizations", "Shopping Centers"]

# Eliminación de muestras
exclude_mask = yelp_business.categories.apply(
    lambda x: any(cat in x for cat in categories_to_exclude)
)
yelp_business = yelp_business[~exclude_mask]

print("Primer filtro")
print("-------------")
print("Muestras eliminadas:", exclude_mask.sum())
print("Restaurantes restantes:", yelp_business.shape[0])

categories = [category for sublist in yelp_business.categories for category in sublist]

unique_categories = set(categories)
print("Categorías únicas (tras primer filtro):", len(unique_categories))

categories = Counter(categories)
categories = categories.most_common()
filename = "categories.txt"
with open(filename, "w", encoding = "utf-8") as f:
    for category, count in categories:
        f.write(f"{category},{count}\n")
print("Categorías almacenadas en:", filename)

## Segundo filtro
# Categorías que marcan casos a excluir del dataset
categories_to_exclude = ["Public Services & Government", "Swimming Pools", "Yelp Events", "Women's Clothing", "Financial Services", "Pet Stores", "Pet Services", "Yoga", "Home Decor", "Departments of Motor Vehicles", "Nail Salons", "Personal Shopping", "Hobby Shops", "Nurseries & Gardening", "Party Equipment Rentals", "Boat Charters", "Party Supplies", "Cosmetics & Beauty Supply", "Banks & Credit Unions", "Jewelry", "Men's Clothing", "Wigs", "Furniture Stores", "Vape Shops"]
low_freq_categories = [cat for cat, count in categories if count < 5]
categories_to_exclude = categories_to_exclude + low_freq_categories
# Eliminación de muestras
exclude_mask = yelp_business.categories.apply(
    lambda x: any(cat in x for cat in categories_to_exclude)
)
yelp_business = yelp_business[~exclude_mask]

# Categorías a eliminar
categories_to_remove = ["Restaurants", "Food", "Event Planning & Services", "Caterers", "Specialty Food", "Venues & Event Spaces", "Shopping", "Party & Event Planning", "Professional Services", "Cooking Schools", "Wholesale Stores", "Drugstores", "Golf Lessons", "Restaurant Supplies"]

# Eliminación de categorías
yelp_business["categories"] = yelp_business["categories"].apply(
    lambda category_list: [
        cat for cat in category_list
        if cat not in categories_to_remove
    ]
).copy()

print("Segundo filtro")
print("-------------")
print("Muestras eliminadas:", exclude_mask.sum())
print("Restaurantes restantes:", yelp_business.shape[0])

categories = [category for sublist in yelp_business.categories for category in sublist]

unique_categories = set(categories)
print("Categorías únicas (tras segundo filtro):", len(unique_categories))

categories = Counter(categories)
categories = categories.most_common()
filename = "categories.txt"
with open(filename, "w", encoding = "utf-8") as f:
    for category, count in categories:
        f.write(f"{category},{count}\n")
print("Categorías almacenadas en:", filename)

#### 2.2. Búsqueda y eliminación de duplicados

print("IDs duplicados:", yelp_business.business_id.duplicated().sum())
print("Restaurantes duplicados (nombre, dirección, ciudad y estado)", yelp_business.loc[: , ["name", "address", "city", "state"]].duplicated().sum())

yelp_business.drop_duplicates(subset = ["name", "address", "city", "state"], inplace=True)
print("Total Restaurants after deleting duplicates:", yelp_business.shape[0])

### 3. Eliminación muestras innecesarias Reviews

#### 3.1. Reviews

yelp_reviews = yelp_reviews[yelp_reviews["business_id"].isin(yelp_business["business_id"])]

print("Total reviews:", yelp_reviews.shape[0])
yelp_reviews.head()

yelp_business.isna().sum()

### 4. Transformación de attributes y categories

#### 4.1. Obtención de todos los atributos únicos y sus posibles valores

##### 4.1.1. Limpieza de la variable

import pandas as pd
import ast

def clean_value(value):
    if isinstance(value, str):
        try:
            posible_dict = ast.literal_eval(value)
            if isinstance(posible_dict, dict):
                return clean_dict(posible_dict)
        except (ValueError, SyntaxError):
            pass
    
        if value.startswith("u'") and value.endswith("'"):
            value = value[2:-1]
        elif value.startswith("'") and value.endswith("'"):
            value = value[1:-1]
        
        value = value.lower().capitalize()
    
    return value

def clean_dict(dic):
    clean_dic = {}
    for k, v in dic.items():
        if isinstance(v, dict):
            clean_dic[k] = clean_dict(v)
        else:
            clean_dic[k] = clean_value(v)
    return clean_dic

yelp_business["attributes"] = yelp_business["attributes"].apply(lambda x: clean_dict(x) if isinstance(x, dict) else x)

##### 4.1.2. Obtención de los valores únicos

from collections import defaultdict

attributes = defaultdict(set)

def get_values(dic, prefix=''):
    for key, value in dic.items():
        complete_key = f"{prefix}.{key}" if prefix else key
        if isinstance(value, dict):
            get_values(value, prefix=complete_key)
        else:
            attributes[complete_key].add(value)

for dic in yelp_business['attributes']:
    if isinstance(dic, dict):
        get_values(dic)

attributes = dict(attributes)

filename = "attributes.txt"
with open(filename, "w", encoding = "utf-8") as f:
    for key, value in attributes.items():
        f.write(f"{key}: {value}\n")
print("Atributos almacenados en:", filename)

### 4.2. To MultiHot

#### 4.2.1. Attributes

def flatten_attributes(dic, prefix=''):
    items = {}
    for k, v in dic.items():
        key = f"{prefix}.{k}" if prefix else k
        if isinstance(v, dict):
            items.update(flatten_attributes(v, prefix=key))
        else:
            items[key] = v
    return items

def clean_value(val):
    val = str(val).lower().strip()
    if val in ['none', 'nan', '']:
        return pd.NA
    return val


flat_attributes = yelp_business["attributes"].apply(
    lambda x: flatten_attributes(x) if isinstance(x, dict) else {}
)


flat_attributes = pd.json_normalize(flat_attributes)


flat_attributes = flat_attributes.map(clean_value)

flat_attributes = pd.get_dummies(flat_attributes, dummy_na=False)

flat_attributes.columns = flat_attributes.columns.str.replace('_true$', '', regex=True)

flat_attributes = flat_attributes.map(lambda x: 1 if x else 0)

flat_attributes = flat_attributes.loc[:, ~flat_attributes.columns.str.endswith('_false')]

print(flat_attributes.shape)
flat_attributes.head()

#### 4.2.2. Categories

from sklearn.preprocessing import MultiLabelBinarizer

mlb = MultiLabelBinarizer()

flat_categories = pd.DataFrame(
    mlb.fit_transform(yelp_business["categories"]),
    columns=mlb.classes_,
    index=yelp_business.index
)

flat_categories.columns = [col.replace(" ", "_") for col in flat_categories.columns]

print(flat_categories.shape)
flat_categories.head()

### 4.3. Attributes text

olivia = Olivia(business_data=yelp_business, review_data = yelp_reviews, attributes=flat_attributes, categories=flat_categories)
olivia.attributes_to_text()
yelp_business = olivia.get_business_data()

### 5. Unión de las estrellas con la reseña

yelp_reviews["text"] = yelp_reviews.apply(
    lambda row: f"Rated {row['stars']} stars out of 5. {row['text']}", axis=1
)

yelp_reviews.head()

### 6. Eliminación de columnas innecesarias

yelp_business.head()

yelp_reviews.head()

yelp_business = yelp_business.drop(columns = ["is_open", "hours", "attributes"])
yelp_reviews = yelp_reviews.drop(columns = ["useful", "funny", "cool", "date"])

print("yelp_business:", yelp_business.shape)
print("yelp_reviews:", yelp_reviews.shape)

### 7. Guardado en csv

yelp_business.to_parquet("../datasets/yelp_business.parquet", index = False)
yelp_reviews.to_parquet("../datasets/yelp_reviews.parquet", index = False)
yelp_photos.to_parquet("../datasets/yelp_photos.parquet", index = False)
flat_attributes.to_parquet("../datasets/attributes_multihot.parquet", index = False)
flat_categories.to_parquet("../datasets/categories_multihot.parquet", index = False)

print("Archivos guardados")
