==> notebooks/05_yelp_validation_preprocessing.ipynb <==
# Procesamiento de los datos para validación

import pandas as pd

## 1. Carga de los datos

yelp_reviews = pd.read_parquet("../datasets/yelp_reviews.parquet")
print("Total reviews:", yelp_reviews.shape[0])
yelp_reviews.head()

## 2. Selección de los usuarios de test

# Selección de reviews positivas
positive_reviews = yelp_reviews[yelp_reviews["stars"] >= 4]
positive_reviews.head()

# Selección de usuarios con 10 o más reviews positivas
eligible_users = positive_reviews.groupby("user_id").size()
eligible_users = eligible_users[eligible_users >= 10].index

# Muestreo
test_users = eligible_users.to_series().sample(n = 10000, random_state = 32)
test_users.head()

### 2.1. Subconjunto de entrenamiento

train_reviews = yelp_reviews[~yelp_reviews["user_id"].isin(test_users)]
print("Total training reviews:", train_reviews.shape[0])
train_reviews.head()

### 2.2. Subconjunto de validación

test_reviews = positive_reviews[positive_reviews["user_id"].isin(test_users)]
print("Total training reviews:", test_reviews.shape[0])
test_reviews.head()

## 3. Selección del restaurante a predecir por usuario

reviews_to_predict = (
    test_reviews
    .groupby("user_id", group_keys=False)
    .apply(lambda df: df.sample(1, random_state=32))["review_id"]
    .tolist()
)
len(reviews_to_predict)

### 3.1. Reseña con restaurante a predecir

business_to_predict = test_reviews[test_reviews["review_id"].isin(reviews_to_predict)]
print("Total business to predict:", business_to_predict.shape[0])
business_to_predict.head()

### 3.2. Reseñas para generar inputs

reviews_to_query = test_reviews[~test_reviews["review_id"].isin(reviews_to_predict)]
print("Total reviews to make queries:", reviews_to_query.shape[0])
reviews_to_query.head()

## 4. Dataset final

reviews_to_query = reviews_to_query.groupby("user_id")["text"].apply(list).reset_index()
validation_df = pd.merge(business_to_predict.loc[: , ["user_id", "business_id"]],
                         reviews_to_query.loc[: , ["user_id", "text"]],
                         how = "left", on = "user_id")

print("Total validation samples:", validation_df.shape[0])
validation_df.head()

validation_df.to_csv("yelp_validation.csv", index = False)
train_reviews.to_csv("yelp_reviews_for_validation.csv", index = False)
